+++ command/command.go
@@ -13,7 +13,13 @@ type projectConfig interface {
 
 type dialog interface {
 	SetMainContaner([]string) (int, string, error)
-	SetStartCommand() (string, error)
+	StartCommand() (string, error)
+	StartDocker() (string, error)
+	DockerService() (string, error)
+}
+
+type containerlist interface {
+	GetContainerList() ([]string, error)
 }
 
 func defineProjectMainContainer(cfg projectConfig, d dialog, containerlist []string) (err error) {