+++ main_docker_test.go
@@ -1,0 +1,17 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestGetContainerList(t *testing.T) {
+	/*dialog := &dialog.Dialog{}
+	dck := &docker.Docker{}
+
+
+	cl := &dockerStartDialog{}
+	cl.setDialog(dialog)
+	cl.setDocker(dck)
+	cl.setDockerService("run docker")*/
+
+}