+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 1.2.0 (28.12.2020)
+- features
+    - 16 Add docker start option
+    - 16 Add docker service dialog for docker starting using jumper
+    - 16 Add unittests for Docker object 25%
+- bugfix
+    - 16 refacored docker instance
+    - update docker API
+
 ## 1.1.0 (27.12.2020)
 - features
     - 14 Added option, which hides long copyright text